Reflect the spec structure more directly.
Add aliases in Library for new functions:
 library_remove_albums +> album_remove_from_library
 library_remove_shows +> shows_remove_from_library
 library_remove_shows +> tracks_remove_from_library
 library_save_album   +> album_save_library
 library_save_shows   +> album_save_library
 library_save_tracks   +> tracks_save_library

Move from Follow to Users:
 follow_check          -> users_check_current_follows
 follow_check_playlist -> users_check_follows_playlist
 follow_get            -> users_get_follows
 follow_get_users      -> users_follow_artists_users
 follow_playlist       -> users_follow_playlist
 unfollow_artists_users-> users_unfollow_artists_users
 unfollow_playlist     -> users_unfollow_playlist

Move from Personalization to Users:
    top_artists         -> users_get_current_user_top_items
    top_tracks          -> users_get_current_user_top_items

modified:   README.md
modified:   docs/src/endpoints.md
modified:   docs/src/index.md
modified:   src/Spotify.jl
modified:   src/by_console_doc/README.md
deleted:    src/by_console_doc/browse.jl
deleted:    src/by_console_doc/follow.jl
modified:   src/by_console_doc/library.jl
deleted:    src/by_console_doc/personalization.jl
modified:   src/by_reference_doc/albums.jl
modified:   src/by_reference_doc/categories.jl
modified:   src/by_reference_doc/playlists.jl
modified:   src/by_reference_doc/shows.jl
modified:   src/by_reference_doc/tracks.jl
modified:   src/by_reference_doc/users.jl
modified:   src/export_structure.jl
modified:   test/calls.jl
renamed:    test/test_browse.jl -> test/test_categories.jl
modified:   test/test_tracks.jl
modified:   test/test_users.jl
