+++ src/by_console_doc/README.md
@@ -2,7 +2,7 @@ Each file in this folder corresponds to ['Console'](https://developer.spotify.co
 
 These partly overlap with ['Reference'](https://developer.spotify.com/documentation/web-api/reference/).
 
-'Browse', 'Follow', 'Library' and 'Personalization' are unique to 'Console'.
+'Browse' and 'Library' are unique to 'Console'.
 
 'Users Profile' describes a subset of 'Reference/Users'.
